feat: stub alerts endpoint and Alert model
